Button now uses Color Enum
