+++ pages/button-test.tsx
@@ -1,0 +1,35 @@
+import * as React from "react";
+import { ActionButton } from "../components/ActionButton";
+import { Color } from "../Color.enum";
+
+const function01 = () => alert("hello 1");
+
+const function02 = () => alert("hello 2");
+
+const ButtonTestPage = () => (
+  <div>
+    <div className="test">WORKING</div>
+
+    <hr></hr>
+
+    <ActionButton
+      thisColor={Color.Primary}
+      label="Primary"
+      clickFunction={function01}
+    ></ActionButton>
+
+    <ActionButton
+      thisColor={Color.Secondary}
+      label="Secondary"
+      clickFunction={function02}
+    ></ActionButton>
+
+    <style jsx>{`
+      .test {
+        color: green;
+      }
+    `}</style>
+  </div>
+);
+
+export default ButtonTestPage;
